# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "excerptlab",
    "contact": {}
  },
  "host": "api.excerptlab.com",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/adminLogin": {
      "post": {
        "summary": "admin login",
        "tags": [
          "admin"
        ],
        "operationId": "adminlogin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/adminloginrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/listUser": {
      "get": {
        "summary": "listUser",
        "tags": [
          "admin"
        ],
        "operationId": "listUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "adminloginrequest": {
      "title": "adminloginrequest",
      "example": {
        "email": "admin@yopmail.com",
        "password": "1234567890"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
...
  },
...
}